Split out more of the tests
